requirements.txt
Flask==3.1.2
tensorflow==2.16.2
numpy==1.26.4
Pillow==12.0.0
requests==2.32.5
